[editor] Remove obsolete `EditorDecorationsService`
`TextEditor.deltaDecorations` should be used instead.
