+++ packages/editor/src/browser/decorations/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright (C) 2018 TypeFox and others.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+export * from './editor-decoration';
+export * from './editor-decoration-style';
+export * from './editor-decorator';
