+++ packages/editor/src/browser/index.ts
@@ -14,4 +14,4 @@ export * from './editor-menu';
 export * from './editor-keybinding-contexts';
 export * from './editor-frontend-module';
 export * from './editor-preferences';
-export * from './editor-decorations-service';
+export * from './decorations';
