+++ packages/monaco/src/browser/monaco-editor.ts
@@ -20,7 +20,6 @@ import {
     TextEditor,
     RevealRangeOptions,
     RevealPositionOptions,
-    EditorDecorationsService,
     DeltaDecorationParams,
     ReplaceTextParams,
 } from '@theia/editor/lib/browser';
@@ -58,7 +57,6 @@ export class MonacoEditor implements TextEditor, IEditorReference {
         readonly node: HTMLElement,
         protected readonly m2p: MonacoToProtocolConverter,
         protected readonly p2m: ProtocolToMonacoConverter,
-        protected readonly decorationsService: EditorDecorationsService,
         options?: MonacoEditor.IOptions,
         override?: IEditorOverrideServices,
     ) {
